Ordered the API and added specific controllers
I have added some modules and updated the mongoose. I highly reccomend to  do npm install again. Also  the calls of the api have changed slightly but they were updated.
