+++ restapi/.env
@@ -1,1 +1,3 @@
 MONGO_URL = mongodb+srv://asw2b:asw2b@lomap.9zrwedt.mongodb.net/?retryWrites=true&w=majority
+API_PORT = 8800
+WEBAPP_PORT = 3000
