+++ restapi/package.json
@@ -2,8 +2,8 @@
   "name": "restapi",
   "version": "1.0.0",
   "scripts": {
-    "start": "ts-node-dev ./server.ts",
-    "dev": "nodemon ./server.ts",
+    "start": "ts-node-dev ./index.ts",
+    "dev": "nodemon ./index.ts",
     "test": "cd tests && jest --runInBand"
   },
   "keywords": [],
