+++ webapp/.env
@@ -1,0 +1,3 @@
+
+API_PORT = 8800
+WEBAPP_PORT = 3000
