+++ webapp/package.json
@@ -15,6 +15,7 @@
     "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
     "axios": "^1.3.4",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "leaflet": "^1.9.3",
     "leaflet-css": "^0.1.0",
