--- webapp/src/api/userApi.ts
@@ -1,21 +1,0 @@
-import { User } from "../shared/shareddtypes";
-
-export async function add(user : User) : Promise<boolean> {
-    const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000/api'
-    let response = await fetch(apiEndPoint+'/users/add', {
-        method: 'POST',
-        headers: {'Content-Type':'application/json'},
-        body: JSON.stringify({'name':user.username, 'email':user.email})
-      });
-    if (response.status===200)
-      return true;
-    else
-      return false;
-}
-
-export async function get() : Promise<User[]> {
-    const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000/api'
-    let response = await fetch(apiEndPoint+'/users/list');
-    //The objects returned by the api are directly convertible to User objects
-    return response.json()
-}
